examples: refactor and move to sidebar
