+++ beaker-vue/src/pages/IntegrationsInterface.vue
@@ -21,6 +21,8 @@
                     :datasources="datasources"
                     :selected-on-load="selectedOnLoad"
                     :folderRoot="datasourcesFolderRoot"
+                    v-model:selected-datasource="selectedDatasource"
+                    v-model:unsaved-changes="unsavedChanges"
                 />
             </div>
         </div>
@@ -72,7 +74,7 @@
                 position="right"
                 highlight="line"
                 :expanded="true"
-                initialWidth="25vi"
+                initialWidth="36vi"
                 :maximized="isMaximized"
             >
                 <SideMenuPanel
@@ -84,6 +86,18 @@
                     <FileContentsPanel
                         :url="previewedFile?.url"
                         :mimetype="previewedFile?.mimetype"
+                    />
+                </SideMenuPanel>
+                <SideMenuPanel
+                    id="examples"
+                    label="Example Editor"
+                    icon="pi pi-file beaker-zoom"
+                    no-overflow
+                >
+                    <ExamplesPanel
+                        v-model="selectedDatasource"
+                        :disabled="!selectedDatasource"
+                        @onchange="unsavedChanges = true"
                     />
                 </SideMenuPanel>
                 <SideMenuPanel id="kernel-logs" label="Logs" icon="pi pi-list" position="bottom">
@@ -116,6 +130,7 @@ import DebugPanel from '../components/panels/DebugPanel.vue'
 
 import DatasourceEditor from '../components/misc/DatasourceEditor.vue';
 import DatasourcePanel from '../components/panels/DatasourcePanel.vue';
+import ExamplesPanel from '../components/panels/ExamplesPanel.vue';
 
 const beakerNotebookRef = ref<BeakerNotebookComponentType>();
 const beakerInterfaceRef = ref();
@@ -165,6 +180,9 @@ const datasources = ref([]);
 const datasourcesFolderRoot = ref("");
 
 const hasOpenedPanelOnce = ref(false);
+
+const selectedDatasource = ref();
+const unsavedChanges = ref<boolean>(false);
 
 type FilePreview = {
     url: string,
@@ -278,8 +296,13 @@ const iopubMessage = (msg) => {
         datasources.value.splice(0, datasources.value.length, ...incomingDatasources);
         datasourcesFolderRoot.value = msg.content.info.datasource_root;
 
+        // rather than onMounted, we fire once we get the message that sets datasources, and hence
+        // creates/populates the panel
         if (!hasOpenedPanelOnce.value) {
-            nextTick(() => sideMenuRef.value.selectPanel('integrations'))
+            nextTick(() => sideMenuRef.value.selectPanel('integrations'));
+            nextTick(() => rightSideMenuRef.value.selectPanel('examples'));
+            (document.querySelector("div.sidemenu.right") as HTMLElement).style.width = '36vi';
+            (document.querySelector("div.sidemenu.left") as HTMLElement).style.width = '25vi';
             hasOpenedPanelOnce.value = true;
         }
     }
@@ -308,8 +331,6 @@ const restartSession = async () => {
     )
     await resetFuture;
 }
-
-
 
 </script>
 
@@ -384,7 +405,7 @@ const restartSession = async () => {
         flex: 1 1000 25vw;
     }
     &.right {
-        flex: 1 1 25vw;
+        flex: 1 1 36vw;
     }
 }
